[REFACTOR/#482] GetAlarmActivity 파일 위치 이동
