+++ app/src/main/java/com/el/yello/presentation/onboarding/activity/OnBoardingActivity.kt
@@ -14,6 +14,7 @@ import com.el.yello.presentation.auth.SignInActivity.Companion.EXTRA_GENDER
 import com.el.yello.presentation.auth.SignInActivity.Companion.EXTRA_KAKAO_ID
 import com.el.yello.presentation.auth.SignInActivity.Companion.EXTRA_NAME
 import com.el.yello.presentation.auth.SignInActivity.Companion.EXTRA_PROFILE_IMAGE
+import com.el.yello.presentation.getalarm.GetAlarmActivity
 import com.el.yello.presentation.onboarding.OnBoardingViewModel
 import com.el.yello.util.extension.yelloSnackbar
 import com.example.ui.base.BindingActivity
